:truck: Move Electron App Outside of Source
