+++ package.json
@@ -7,12 +7,13 @@
     "email": "hello@process-engine.io",
     "url": "http://process-engine.io"
   },
-  "main": "electron_app/electron.js",
+  "main": "dist/electron_app/electron_app/electron.js",
   "scripts": {
     "start": "./bin/bpmn-studio.js",
     "postinstall": "npm rebuild",
     "start_dev": "au run --watch",
-    "build": "au build",
+    "build-electron-stuff": "tsc electron_app/electron.ts --outDir 'dist/electron_app' --target es5 --lib esnext --lib dom --esModuleInterop true --resolveJsonModule",
+    "build": "npm run build-electron-stuff && au build",
     "reset": "bash dev_tooling/scripts/reset_setup.bash",
     "reinstall": "OLD_PATH=$PATH bash dev_tooling/scripts/reinstall_setup.bash && unset OLD_PATH",
     "reinstall-clean": "OLD_PATH=$PATH bash dev_tooling/scripts/reinstall_clean_setup.bash && unset OLD_PATH",
@@ -26,7 +27,7 @@
     "electron-build-macos": "npm run electron-prepare && electron-builder --config build/electron-builder.yml --publish never --macos && npm run electron-cleanup",
     "electron-cleanup": "node electron-builder/reset-name-in-package-json.js",
     "electron-prepare": "electron-builder install-app-deps && node electron-builder/set-name-in-package-json.js && electron-rebuild && npm run build && node electron-builder/prepare-electron-build.js",
-    "electron-start-dev": "npm run build && electron electron_app/electron.js",
+    "electron-start-dev": "npm run build && electron dist/electron_app/electron_app/electron.js",
     "jenkins-electron-install-app-deps": "electron-builder install-app-deps",
     "jenkins-electron-rebuild-native": "electron-rebuild --force",
     "jenkins-electron-build-windows": "electron-builder --publish never --windows",
